+++ test/tokens/Erc2612.sol
@@ -24,7 +24,7 @@ contract Erc2612 is PErc20, IErc2612 {
   /// @param d number of decimals for the token
   /// @param n name for the token
   /// @param s symbol for the token
-  constructor(uint16 d, string memory n, string memory s) PErc20(d, n, s) {
+  constructor(uint8 d, string memory n, string memory s) PErc20(d, n, s) {
     DOMAIN = Hash.domain(n, '1', block.chainid, address(this));
   }
